requirements.txt
aiomax>=1.0.0
aiohttp
aiofiles
python-dotenv